BOSW1510437 insite - opschonen vge adressen - Added echo of actions

+++ api/v3/VgeAdres/RemoveVgeHovEnded.php
@@ -21,11 +21,18 @@ function _civicrm_api3_vge_adres_removevgehovended_spec(&$spec) {
  * @see civicrm_api3_create_error
  * @throws API_Exception
  */
-function civicrm_api3_vge_adres_removevgehovended($params) {
+function civicrm_api3_vge_adres_removevgehovended($params) {  
+  $debug = CRM_Utils_Array::value('debug', $params, false);
+    
   $return['is_error'] = false;
   
+  if($debug){
+    $return['message'][] = ts('Debug is on !');
+    echo ts('Debug is on !') . '<br/>' . PHP_EOL;
+  }
+  
   // households
-  $return_household = vge_adres_removevgehovended_hov_household();
+  $return_household = vge_adres_removevgehovended_hov_household($debug);
   if($return_household['is_error']){
     $return = $return_household;
     $return['is_error'] = true;
@@ -33,7 +40,7 @@ function civicrm_api3_vge_adres_removevgehovended($params) {
   }
   
   // organizations
-  $return_organization = vge_adres_removevgehovended_hov_organization();
+  $return_organization = vge_adres_removevgehovended_hov_organization($debug);
   if($return_organization['is_error']){
     $return = array_merge($return_household, $return_organization);
     $return['is_error'] = true;
@@ -42,16 +49,21 @@ function civicrm_api3_vge_adres_removevgehovended($params) {
   
   $return = array_merge($return_household, $return_organization);
   
+  if($debug){
+    CRM_Utils_System::civiExit();
+  }
+  
   return civicrm_api3_create_success($return);
 }
 
-function vge_adres_removevgehovended_hov_household(){ 
+function vge_adres_removevgehovended_hov_household($debug){ 
   $return['is_error'] = false;
   $return['message'] = [];
   
   $locationtype = CRM_Utils_DgwUtils::getLocationType(['name' => 'VGEadres'], 'getsingle');
   if(!$locationtype){
     $return['error_message'] = sprintf('Cannot get the location_type_id of the VGE adres');
+    echo sprintf('Cannot get the location_type_id of the VGE adres') . '<br/>' . PHP_EOL;
     $return['is_error'] = true;
     return $return;
   }
@@ -59,13 +71,15 @@ function vge_adres_removevgehovended_hov_household(){
   $customgroup = CRM_Utils_DgwUtils::getCustomGroup(['name' => 'Huurovereenkomst (huishouden)'], 'getsingle'); 
   if(!$customgroup){
     $return['error_message'] = sprintf('Cannot get the custom_group_id of the Huurovereenkomst (huishouden)');
+    echo sprintf('Cannot get the custom_group_id of the Huurovereenkomst (huishouden)') . '<br/>' . PHP_EOL;
     $return['is_error'] = true;
     return $return;
   }
   
   $customfield = CRM_Utils_DgwUtils::getCustomFieldAll(['custom_group_id' => $customgroup['id']]);  
   if(!$customfield){
-    $return['error_message'] = sprintf('Cannot get all the custom fields of the Huurovereenkomst (huishouden)', mysql_error($link));
+    $return['error_message'] = sprintf('Cannot get all the custom fields of the Huurovereenkomst (huishouden)');
+    echo sprintf('Cannot get all the custom fields of the Huurovereenkomst (huishouden)') . '<br/>' . PHP_EOL;
     $return['is_error'] = true;
     return $return;
   }
@@ -94,6 +108,9 @@ function vge_adres_removevgehovended_hov_household(){
   $dao = CRM_Core_DAO::executeQuery($query, $params);
     
   while ($dao->fetch()) {    
+    //$return['message'][] = ts('Next household with id \'' . $dao->id);
+    echo ts('Next household with id \'' . $dao->id) . '<br/>' . PHP_EOL;
+      
     /**
      * Sometimes the address of the HOV is like Adriaan Pauwstraat,15,7331 NH,Apeldoorn
      * We have to convert it to Adriaan Pauwstraat 15 to compare it with the street address
@@ -110,33 +127,46 @@ function vge_adres_removevgehovended_hov_household(){
     foreach($addresses['values'] as $address){
       // delete if household vge_adres_first_7 is the same as the hoofdhuurder street_address
       if($vge_adres == $address['street_address']){
-        CRM_Utils_DgwUtils::deleteAddress($address['id']);
         $return['message'][] = ts('Household with id \'' . $dao->id . '\', the address with street address \'' . $address['street_address'] . '\' with address id \'' . $address['id'] . '\', the address is deleted !');
+        echo ts('Household with id \'' . $dao->id . '\', the address with street address \'' . $address['street_address'] . '\' with address id \'' . $address['id'] . '\', the address is deleted !') . '<br/>' . PHP_EOL;
+        if(!$debug){
+          CRM_Utils_DgwUtils::deleteAddress($address['id']);
+        }
       }
     }
     
     // get hoofdhuurders
     $hoofdhuurders = CRM_Utils_DgwUtils::getHoofdhuurders($dao->id, false);
     foreach($hoofdhuurders as $hoofdhuurder){
+      //$return['message'][] = ts('Next hoofdhuurder of household with id \'' . $dao->id . '\' the with id \'' . $hoofdhuurder['contact_id']);
+      echo ts('Next hoofdhuurder of household with id \'' . $dao->id . '\' the with id \'' . $hoofdhuurder['contact_id']);
       $addresses = CRM_Utils_DgwUtils::getAddress(['contact_id' => $hoofdhuurder['contact_id'], 'location_type_id' => $locationtype['id']], 'get');
       foreach($addresses['values'] as $address){
         // delete if household vge_adres_first_7 is the same as the hoofdhuurder street_address
         if($vge_adres == $address['street_address']){
-          CRM_Utils_DgwUtils::deleteAddress($address['id']);
           $return['message'][] = ts('Household with id \'' . $dao->id . '\', there the hoofdhuurder with id \'' . $hoofdhuurder['contact_id'] . '\' the address with street address \'' . $address['street_address'] . '\' with address id \'' . $address['id'] . '\', the address is deleted !');
+          echo ts('Household with id \'' . $dao->id . '\', there the hoofdhuurder with id \'' . $hoofdhuurder['contact_id'] . '\' the address with street address \'' . $address['street_address'] . '\' with address id \'' . $address['id'] . '\', the address is deleted !') . '<br/>' . PHP_EOL;
+          if(!$debug){
+            CRM_Utils_DgwUtils::deleteAddress($address['id']);
+          }
         }
       }
     }
     
     // get medehuurders
     $medehuurders = CRM_Utils_DgwUtils::getMedeHuurders($dao->id, false);
-    foreach($medehuurders as $medehuurder){      
+    foreach($medehuurders as $medehuurder){     
+      //$return['message'][] = ts('Next medehuurder of household with id \'' . $dao->id . '\' the with id \'' . $medehuurder['medehuurder_id']);
+      echo ts('Next medehuurder of household with id \'' . $dao->id . '\' the with id \'' . $medehuurder['medehuurder_id']);
       $addresses = CRM_Utils_DgwUtils::getAddress(['contact_id' => $medehuurder['medehuurder_id'], 'location_type_id' => $locationtype['id']], 'get');
       foreach($addresses['values'] as $address){
         // delete if household vge_adres_first_7 is the same as the medehuurder street_address
         if($vge_adres == $address['street_address']){
-          CRM_Utils_DgwUtils::deleteAddress($address['id']);
           $return['message'][] = ts('Household with id \'' . $dao->id . '\', there the medehuurder with id \'' . $medehuurder['medehuurder_id'] . '\' the address with street address \'' . $address['street_address'] . '\' with address id \'' . $address['id'] . '\', the address is deleted !');
+          echo ts('Household with id \'' . $dao->id . '\', there the medehuurder with id \'' . $medehuurder['medehuurder_id'] . '\' the address with street address \'' . $address['street_address'] . '\' with address id \'' . $address['id'] . '\', the address is deleted !') . '<br/>' . PHP_EOL;
+          if(!$debug){
+            CRM_Utils_DgwUtils::deleteAddress($address['id']);
+          }
         }
       }
     }
@@ -145,13 +175,14 @@ function vge_adres_removevgehovended_hov_household(){
   return $return;
 }
 
-function vge_adres_removevgehovended_hov_organization(){
+function vge_adres_removevgehovended_hov_organization($debug){
   $return['is_error'] = false;
   $return['message'] = [];
   
   $locationtype = CRM_Utils_DgwUtils::getLocationType(['name' => 'VGEadres'], 'getsingle');
   if(!$locationtype){
     $return['error_message'] = sprintf('Cannot get the location_type_id of the VGE adres');
+    echo sprintf('Cannot get the location_type_id of the VGE adres') . '<br/>' . PHP_EOL;
     $return['is_error'] = true;
     return $return;
   }
@@ -159,13 +190,15 @@ function vge_adres_removevgehovended_hov_organization(){
   $customgroup = CRM_Utils_DgwUtils::getCustomGroup(['name' => 'Huurovereenkomst (organisatie)'], 'getsingle'); 
   if(!$customgroup){
     $return['error_message'] = sprintf('Cannot get the custom_group_id of the Huurovereenkomst (organistatie)');
+    echo sprintf('Cannot get the custom_group_id of the Huurovereenkomst (organistatie)') . '<br/>' . PHP_EOL;
     $return['is_error'] = true;
     return $return;
   }
   
   $customfield = CRM_Utils_DgwUtils::getCustomFieldAll(['custom_group_id' => $customgroup['id']]);
   if(!$customfield){
-    $return['error_message'] = sprintf('Cannot get all the custom fields of the Huurovereenkomst (organistatie)', mysql_error($link));
+    $return['error_message'] = sprintf('Cannot get all the custom fields of the Huurovereenkomst (organistatie)');
+    echo sprintf('Cannot get all the custom fields of the Huurovereenkomst (organistatie)') . '<br/>' . PHP_EOL;
     $return['is_error'] = true;
     return $return;
   }
@@ -193,7 +226,10 @@ function vge_adres_removevgehovended_hov_organization(){
   );
   $dao = CRM_Core_DAO::executeQuery($query, $params);
     
-  while ($dao->fetch()) {    
+  while ($dao->fetch()) {
+    //$return['message'][] = ts('Next organization with id \'' . $dao->id);
+    echo ts('Next organization with id \'' . $dao->id) . '<br/>' . PHP_EOL;
+    
     /**
      * Sometimes the address of the HOV is like Adriaan Pauwstraat,15,7331 NH,Apeldoorn
      * We have to convert it to Adriaan Pauwstraat 15 to compare it with the street address
@@ -210,8 +246,11 @@ function vge_adres_removevgehovended_hov_organization(){
     foreach($addresses['values'] as $address){
       // delete if household vge_adres_first_7 is the same as the hoofdhuurder street_address
       if($vge_adres[0] == $address['street_address']){
-        CRM_Utils_DgwUtils::deleteAddress($address['id']);
         $return['message'][] = ts('Organization with id \'' . $dao->id . '\', the address with street address \'' . $address['street_address'] . '\' with address id \'' . $address['id'] . '\', the address is deleted !');
+        echo ts('Organization with id \'' . $dao->id . '\', the address with street address \'' . $address['street_address'] . '\' with address id \'' . $address['id'] . '\', the address is deleted !') . '<br/>' . PHP_EOL;
+        if(!$debug){
+          CRM_Utils_DgwUtils::deleteAddress($address['id']);
+        }
       }
     }
   }
